Clarify bakers math
The previous text did not make it clear that bakers math specifies percentages as a fraction of the amount of flour, not as a fraction of the final weight.  Updated wording clarifies this.

Also proposed some grammar changes.  (The number of variables impacting a pizza is a discrete integer, and not a continuous number, and so "fewer" is used in place of "less".  En-dashes are used in sentences, not minus signs.)